Merge branch 'master' into pahender/DAOS-13536-simple

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium-verbs-provider: true
Skip-func-hw-test-large: true
Skip-func-hw-test-medium-ucx-provider: false
Allow-unstable-test: true
Test-tag: vm test_dmg_network_scan_basic NetworkFailureTest + ReplayTests

+++ src/tests/ftest/util/fault_config_utils.py
@@ -303,6 +303,9 @@ class FaultInjection():
         Returns:
            error_list (list) : Errors during removing fault files (if any).
         """
+        if not self.fault_file:
+            return []
+
         # Remove the fault injection files on the hosts.
         error_list = []
         command = "rm -f {}".format(self.fault_file)
